+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+## [0.45.0](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.45.0) - 2021-10-13
+
+**Features**
+
+- add `TreeView` component accessors to programmatically expand/collapse nodes
+
+## [0.44.7](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.44.7) - 2021-10-13
+
+**Fixes**
+
+- add missing `value` prop to `Checkbox`
+
+## [0.44.6](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.44.6) - 2021-10-11
+
+**Fixes**
+
+- fix `DataTable` sorting to tolerate `null`, `undefined` values
+
 ## [0.44.5](https://github.com/carbon-design-system/carbon-components-svelte/releases/tag/v0.44.5) - 2021-10-04
 
 **Fixes**